Add tty_bsd.go for Mac OS X/FreeBSD
