+++ tty_bsd.go
@@ -1,0 +1,22 @@
+// +build darwin freebsd
+
+package main
+
+import (
+  "os"
+  "syscall"
+  "unsafe"
+)
+
+func isatty() bool {
+  var termios syscall.Termios
+  _, _, err := syscall.Syscall6(syscall.SYS_IOCTL, os.Stdin.Fd(), uintptr(syscall.TIOCGETA), uintptr(unsafe.Pointer(&termios)), 0, 0, 0)
+  return err == 0
+}
+
+func tty_ready() error {
+  return nil
+}
+
+func tty_term() {
+}
